requirements.txt
PyJWT~=2.8.0
cryptography==41.0.3
python-dotenv==1.0.0
requests~=2.31.0
flask~=2.3.3
Jinja2~=3.1.2
WTForms~=3.0.1
Flask-WTF~=1.1.1